database/file_manager/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
python-multipart